test_missing_obj_finder: fix bogus haves test
The code under test states in a comment that non-local haves are legal,
but that non-local wants are not. The test was attempting to enforce
that both non-local haves and wants were not legal, but was getting
lucky due to an error in the test implementation.

The bogus wants test is fixed in the following commit.