allow for reuse of existing deltas while creating pack files
This feature allows the ``dulwich pack-objects`` command to
complete in a somewhat reasonable amount of time on small
repositories, provided a pack file which can provide deltas
already exists.

Reusing deltas from an existing pack file avoids deltification
overhead while creating a new pack file. This optimization is
especially important when serving clones and fetches from a server.

For testing purposes, add two options to ``dulwich pack-objects``:
  --deltify: enables deltification and delta-reuse
  --no-reuse-deltas: disables delta-reuse during deltification

The write_pack_objects() now accepts an optional Pack file object
argument which enables reuse of deltas.

If deltification is enabled and delta-reuse is not explicitly disabled,
the pack_objects() routine in porcelain.py picks a suitable pack file
for reuse, which is the pack file which contains the largest amount of
objects. A delta found in this pack file can be reused if its delta-base
will also be part of the generated pack file.

There are some limitations which should be lifted eventually:

Reused deltas are stored as ref-deltas only. There is no logic
yet to figure out where the base of a reused offset-delta has
been placed in the new pack file.

We decompress deltas for reuse and compress them again when writing
them to the new pack file. More code refactoring would be needed to
support direct copying from the source pack file in order to avoid
re-compression overhead.